add int-bot and the-collector to a single cargo workspace
I have been developing int-bot and the-collector in separate cargo
projects. This has shown some issues in development, as the two
applications are very tightly coupled at this moment in time, meaning
there is a lot of shared logic between them. By having them in the same
workspace, it can be easier to set them up to share libraries (that are
also in the workspace).