use summoner create_time if no latest match yet
For a long time, including in the first iteration of the int bot a few
years ago, the behavior when adding a new summoner to track was to
retrieve their last game to have a reference point to start with. We can
instead reference when the summoner was added to the database if there
is no game for them in the database yet.