fix 'where in' queries
sqlx does not have great support for WHERE IN clauses, and it turns out
the original implementation done here did not work. In order to be able
to safely construct there WHERE IN query, we mostly have to manually
build the parameterized query. This new implementation also loses the
effect of having compile-time checks for these queries (get_leaderboard
and get_matches).